Done: move from int to long long

+++ GeneralAlgebraTeamLabProject/Lib/task5/pollardfactorization.cpp
@@ -5,38 +5,38 @@
 #include <stdexcept>
 
 // finds greatest common divider
-static int gcd(int a, int b) {
+static long long gcd(long long a, long long b) {
     return b == 0 ? a : gcd (b, a % b);
 }
 
-static bool isPrime(int a) {
-    for (int i = 2; i <= sqrt(a); ++i) {
+static bool isPrime(long long a) {
+    for (long long i = 2; i <= sqrt(a); ++i) {
         if (a % i == 0) return false;
     }
     return true;
 }
 
-static int F(int x, int a, int mod) {
+static long long F(long long x, long long a, long long mod) {
     return (x * x + (1 + 2 * a)) % mod;
 }
 
-static std::vector<int> rhoPollard(int N) {
+static std::vector<long long> rhoPollard(long long N) {
     if (N < 1) {
         throw std::invalid_argument(N + " is less than 1");
     }
     if (N == 1) {
-        return std::vector<int>{1};
+        return std::vector<long long>{1};
     }
     if (N % 2 == 0) {
-        std::vector<int> t = rhoPollard(N / 2);
+        std::vector<long long> t = rhoPollard(N / 2);
         t.push_back(2);
         return t;
     }
-    std::vector<int> ans;
-    int x = 2;
-    int y = 2;
-    int gcd1;
-    int a = rand() % 10;
+    std::vector<long long> ans;
+    long long x = 2;
+    long long y = 2;
+    long long gcd1;
+    long long a = rand() % 10;
     do {
         x = F(x, a, N);
         y = F(F(y, a, N), a, N);
@@ -46,23 +46,23 @@ static std::vector<int> rhoPollard(int N) {
     if (isPrime(gcd1)) {
         ans.push_back(gcd1);
     } else {
-        std::vector<int>  t = rhoPollard(gcd1);
+        std::vector<long long>  t = rhoPollard(gcd1);
         ans.insert(ans.end(), t.begin(), t.end());
     }
 
-    int zvor = N / gcd1;
+    long long zvor = N / gcd1;
     if (isPrime(zvor)) {
         ans.push_back(zvor);
     } else {
-        std::vector<int>  t = rhoPollard(zvor);
+        std::vector<long long>  t = rhoPollard(zvor);
         ans.insert(ans.end(), t.begin(), t.end());
     }
 
     return ans;
 }
 
-std::vector<int> PollardFactorization::factorize(int number) {
-    std::vector<int> primeNumbers = rhoPollard(number);
+std::vector<long long> PollardFactorization::factorize(long long number) {
+    std::vector<long long> primeNumbers = rhoPollard(number);
     sort(primeNumbers.begin(), primeNumbers.end());
     return primeNumbers;
 }
